Rename app to flask_socket_generator_app
