+++ tests/test_project_manager.py
@@ -1,8 +1,8 @@
 import os
 import shutil
 import pytest
-from app.project_manager import ProjectManager
-from app.project_builder import ProjectBuilder
+from flask_socket_generator_app.project_manager import ProjectManager
+from flask_socket_generator_app.project_builder import ProjectBuilder
 
 
 __BASE_DIRECTORY = "tests/"
